Added a pin button at the top, Updated event
Hopefully that button works! I copied the style from the one on the right. 

Also I commented out the links to the Eventbrite page since there isn't a location yet... I'm so sorry I know that's not the way things are done. Just want to get this up so we can start promoting it and selling pins!